removed Prev / Next tool tips when no more pages (#456)
Tool-tip is cropped after setting max-width: {X}em and padding < 3em in css, so I did this to remove it.

Another option is to shift the dissplay of the individual tooltips:

'left: 50%;' --> 'left: 130%;' for Prev
'left: 50%;' --> 'left:   ?%;' for Next

Prev and Next would need individual designations, at the moment the css changes all tooltips, so I don't think this is viable in the present config.

Not sure if you want this hack?

Anyway, I'm learning how to use this.

Co-authored-by: laz <laz@vltra.plus>
Reviewed-on: https://git.mills.io/yarnsocial/yarn/pulls/456
Reviewed-by: James Mills <james@mills.io>
Co-authored-by: laz <laz@noreply@mills.io>
Co-committed-by: laz <laz@noreply@mills.io>