Add a class to the twt navigation (#528)
Add a class to the twt navigation (edit, delete, etc.), so that changes to it do not collide with other <nav>. If there is a better way to do this, I am all ears! :-) A CSS change to apply a smaller font to the links within that <nav> comes next.

Reviewed-on: https://git.mills.io/yarnsocial/yarn/pulls/528
Co-authored-by: I am Fastidious <fastidious@noreply@mills.io>
Co-committed-by: I am Fastidious <fastidious@noreply@mills.io>