feeds: expand external feeds view, add placeholder avatars (#573)
The `/feeds` view was expanded with additional information, and for external feeds without avatars, placeholder avatars are generated.

NOTE: I tested it and it appears to work fine, with the exception of an odd scaling issue on wider resolutions where the CSS appears to break. I'm not sure what the source of the issue is.

Co-authored-by: Nate <servusdei@programmer.net>
Reviewed-on: https://git.mills.io/yarnsocial/yarn/pulls/573
Co-authored-by: servusdei2018 <servusdei2018@noreply@mills.io>
Co-committed-by: servusdei2018 <servusdei2018@noreply@mills.io>