CSS adjustments for mobile/responsive view (#572)
I swear to all that is holy, this better pass drone testing...

Okay, so, this is to fix all the lovely fuck-ups.  I hope that this works for everybody.  Like I said I checked and re-checked quite a lot.

*What could go wrong...*

Added in class references for top navigation to adjust icons.

NO INLINE CSS. Yes yes, none of those shenanigans.

Co-authored-by: ullarah <phil@git.ullarah.com>
Reviewed-on: https://git.mills.io/yarnsocial/yarn/pulls/572
Co-authored-by: ullarah <ullarah@noreply@mills.io>
Co-committed-by: ullarah <ullarah@noreply@mills.io>