Fix #602: Referrer not remembered after login (#614)
This commit should fix issue #602. Referer policy was added to translations' login links, and on the login page, the referer is set as the document root, ensuring it gets passed to the login post handler. Previously, `https://www.twtxt.net/login` had been sent as the referer, because there was a double referal going on: `actual page -> GET login -> POST login`, so the post handler would think that the referer was always the login page, and not the actual page.

Co-authored-by: Nate <servusdei@programmer.net>
Reviewed-on: https://git.mills.io/yarnsocial/yarn/pulls/614
Co-authored-by: servusdei2018 <servusdei2018@noreply@mills.io>
Co-committed-by: servusdei2018 <servusdei2018@noreply@mills.io>