Pico.css update to current version (#625)
Fixed a lot of weirdness with latest pico.css in yarn css file.

Glaring issue with responsive view with the burger menu. Still figuring it out but wanted to make a PR with current fixes.

Also using quite a lot of !important tags in yarn.css which I'm not happy about.

@prologic @fastidious @markwylde

Can see update on instance: https://txt.quisquiliae.com

Co-authored-by: Phil S <phil@git.ullarah.com>
Co-authored-by: James Mills <prologic@shortcircuit.net.au>
Reviewed-on: https://git.mills.io/yarnsocial/yarn/pulls/625
Co-authored-by: Phil S. <ullarah@noreply@mills.io>
Co-committed-by: Phil S. <ullarah@noreply@mills.io>