Redo of how image handlers deal with modal and external images (#668)
Going to continue this conversation here. @prologic @fastidious

I assume this 'zooming' thing is what is referred to as cropping.  The image is center-cropped as to reduce the amount of scrolling a user will require if somebody uploads a large vertical image.

The current 'master' was not currently dealing with external images correctly.  I am hoping that this PR will manage to fix this issue.

**Alternative Idea:**
Instead of a modal box, we still center-crop the image, but on click/tap the image enlarges (and shrinks) instead of appearing on a floating box. However this essentially reverts back to loading the full image.

Co-authored-by: Phil S <phil@git.ullarah.com>
Co-authored-by: James Mills <prologic@shortcircuit.net.au>
Reviewed-on: https://git.mills.io/yarnsocial/yarn/pulls/668
Co-authored-by: Phil S. <ullarah@noreply@mills.io>
Co-committed-by: Phil S. <ullarah@noreply@mills.io>